inventory/requirements.txt
Python-decouple==3.8
Django==3.2.25
djangorestframework==3.12.4
psycopg2-binary==2.9.10